requirements.txt
annotated-types==0.7.0
arrow==1.2.3
attrs==23.1.0
blinker==1.9.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
ics==0.7.2
idna==3.4
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
pyairtable==3.0.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.8.2
python-dotenv==0.21.1
python-slugify==8.0.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
TatSu==5.8.3
text-unidecode==1.3
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==1.26.14
Werkzeug==3.1.3
